Implement suggested changes
