+++ lib/generators/rolemodel/testing/jasmine_playwright/README.md
