+++ lib/generators/rolemodel/testing/jasmine_playwright/templates/example_spec.js
