+++ lib/generators/rolemodel/testing/jasmine_playwright/templates/jp-runner.config.mjs
