+++ lib/generators/rolemodel/testing/jasmine_playwright/templates/support/react.mjs
