+++ lib/generators/rolemodel/testing/rspec/rspec_generator.rb
@@ -17,9 +17,9 @@ module Rolemodel
         gem_group :test do
           gem 'capybara'
           gem 'capybara-playwright-driver'
-          gem 'webdrivers'
           gem 'marsh_grass'
           gem 'pry'
+          gem 'webdrivers'
         end
         run_bundle
       end
@@ -38,10 +38,10 @@ module Rolemodel
         template 'support/capybara_drivers.rb', 'spec/support/capybara_drivers.rb'
         template 'support/capybara_testid.rb', 'spec/support/capybara_testid.rb'
         template 'support/helpers/action_cable_helper.rb', 'spec/support/helpers/action_cable_helper.rb'
+        template 'support/helpers/capybara_helper.rb', 'spec/support/helpers/capybara_helper.rb'
+        template 'support/helpers/playwright_helper.rb', 'spec/support/helpers/playwright_helper.rb'
         template 'support/helpers/select_helper.rb', 'spec/support/helpers/select_helper.rb'
         template 'support/helpers/test_element_helper.rb', 'spec/support/helpers/test_element_helper.rb'
-        template 'support/helpers/playwright_helper.rb', 'spec/support/helpers/playwright_helper.rb'
-        template 'support/helpers/capybara_helper.rb', 'spec/support/helpers/capybara_helper.rb'
         template 'support/helpers.rb', 'spec/support/helpers.rb'
         append_file '.gitignore', 'spec/examples.txt'
       end
