+++ spec/generators/rolemodel/rspec_generator_spec.rb
@@ -10,9 +10,9 @@ RSpec.describe Rolemodel::Testing::RspecGenerator, type: :generator do
 
   it 'adds the correct helpers' do
     assert_file 'spec/support/helpers/action_cable_helper.rb'
+    assert_file 'spec/support/helpers/capybara_helper.rb'
+    assert_file 'spec/support/helpers/playwright_helper.rb'
     assert_file 'spec/support/helpers/select_helper.rb'
     assert_file 'spec/support/helpers/test_element_helper.rb'
-    assert_file 'spec/support/helpers/playwright_helper.rb'
-    assert_file 'spec/support/helpers/capybara_helper.rb'
   end
 end
